requirements.txt
crewai
pydantic
streamlit
python-dotenv
pylint
